Refactor and add trainer
