--- climatenet/models/cgnet/cgnet.json
@@ -1,32 +1,0 @@
-{
-    "architecture": "cgnet",
-    "seed": 42,
-    "lr": 0.001,
-    "train_batch_size": 4,
-    "pred_batch_size": 8,
-    "epochs": 15,
-    "fields": {
-        "TMQ": {
-            "mean": 19.21859,
-            "std": 15.81723
-        },
-        "U850": {
-            "mean": 1.55302,
-            "std": 8.29764
-        },
-        "V850": {
-            "mean": 0.25413,
-            "std": 6.23163
-        },
-        "PSL": {
-            "mean": 100814.414,
-            "std": 1461.2227
-        }
-    },
-    "labels": [
-      "Background",
-      "Tropical Cyclone",
-      "Atmospheric River"
-    ],
-    "description": "The basic CGNet model. You can use this config to train your own model, or load it with our trained weights."
-}
